refactor: reorganize into metadata and content
